requirements.txt
absl-py==0.14.1
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
blis==0.7.4
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cymem==2.0.5
de-core-news-sm==3.1.0
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
dotmap==1.3.24
en-core-web-sm==3.1.0
entrypoints==0.3
google-auth==2.3.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
idna==3.3
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
testpath==0.5.0
thinc==8.0.10
torch==1.9.1+cu111
torchaudio==0.9.1
torchtext==0.10.1
torchvision==0.10.1+cu111
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
yacs==0.1.8
